Search-Function-Dev-Branch
This dev branch adds some new code to index.js and some of the models for case-insensitivity. And then I added search.js, search.jsx, and search.css to meet the requirements for the search function.

I had to modify some other files to bypass authentication to test the search function. That should not be included in this PR so please let me know if I need to resubmit a new one due to that