Brian-Test-Branch-Issue-61
This branch updates log.jsx, log.css, search.jsx, and search.css. I had to tweak some of the logic to correct unwanted behavior using the search function and I added a 'Clear' button so searches can be cleared. I tweaked some of the styling for both pages as well to improve consistency.