Issue98 and some UI/UX improvements
This issue ties the recipe modal to the log button on recipe cards in the search page. I also changed the header for the recipe modal because it previously felt repetitive pressing the log button twice and by changing the header to "Select recipe and servings:", it doesn't feel repetitive anymore. I also made it so that logged recipes will display their servings amount